[기능개선][관리자] 관리자 api AdminService 통일 feat: 공연정보, 공연장 정보 저장 메서드 adminService 추가 #103
