+++ src/main/java/com/ticketmate/backend/controller/admin/AdminController.java
@@ -1,15 +1,13 @@
 package com.ticketmate.backend.controller.admin;
 
 import com.ticketmate.backend.controller.admin.docs.AdminControllerDocs;
-import com.ticketmate.backend.object.dto.auth.request.CustomOAuth2User;
-import com.ticketmate.backend.object.dto.concert.request.ConcertInfoRequest;
-import com.ticketmate.backend.object.dto.concerthall.request.ConcertHallInfoRequest;
 import com.ticketmate.backend.object.dto.admin.request.PortfolioSearchRequest;
 import com.ticketmate.backend.object.dto.admin.response.PortfolioForAdminResponse;
 import com.ticketmate.backend.object.dto.admin.response.PortfolioListForAdminResponse;
+import com.ticketmate.backend.object.dto.auth.request.CustomOAuth2User;
+import com.ticketmate.backend.object.dto.concert.request.ConcertInfoRequest;
+import com.ticketmate.backend.object.dto.concerthall.request.ConcertHallInfoRequest;
 import com.ticketmate.backend.service.admin.AdminService;
-import com.ticketmate.backend.service.concert.ConcertService;
-import com.ticketmate.backend.service.concerthall.ConcertHallService;
 import com.ticketmate.backend.util.log.LogMonitoringInvocation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -31,8 +29,6 @@ import java.util.UUID;
 )
 public class AdminController implements AdminControllerDocs {
 
-    private final ConcertService concertService;
-    private final ConcertHallService concertHallService;
     private final AdminService adminService;
 
     @Override
@@ -41,7 +37,7 @@ public class AdminController implements AdminControllerDocs {
     public ResponseEntity<Void> saveHallInfo(
             @AuthenticationPrincipal CustomOAuth2User customOAuth2User,
             @Valid @RequestBody ConcertHallInfoRequest request) {
-        concertHallService.saveHallInfo(request);
+        adminService.saveHallInfo(request);
         return ResponseEntity.ok().build();
     }
 
@@ -51,7 +47,7 @@ public class AdminController implements AdminControllerDocs {
     public ResponseEntity<Void> saveConcertInfo(
             @AuthenticationPrincipal CustomOAuth2User customOAuth2User,
             @Valid @ModelAttribute ConcertInfoRequest request) {
-        concertService.saveConcertInfo(request);
+        adminService.saveConcertInfo(request);
         return ResponseEntity.ok().build();
     }
